fix broken storybook stories - harbor list - customer list - page - harbor details - individual harbor page (no content) - individual customer page (no content)
