+++ src/domain/customers/CustomerListComponent.stories.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { HashRouter } from 'react-router-dom';
 
 import CustomerListComponent from './CustomerListComponent';
 import CustomersPage from './CustomersPage';
@@ -7,6 +8,7 @@ import dummyCustomers from './__mocks__/data.json';
 
 export default {
   component: CustomerListComponent,
+  decorators: [storyFn => <HashRouter>{storyFn()}</HashRouter>],
   title: 'CustomerList',
 };
