+++ src/domain/page/Page.tsx
@@ -4,7 +4,7 @@ import { useTranslation } from 'react-i18next';
 
 import Layout from '../../common/layout/Layout';
 import Sidebar from '../../common/sidebar/Sidebar';
-import { IconApplications, IconBoat, IconFemale } from '../../common/icons';
+import { IconApplications, IconBoat, IconPerson } from '../../common/icons';
 import InternalLink from '../../common/internalLink/NavLink';
 import PageHeader from '../pageHeader/PageHeaderContainer';
 import styles from '../../common/icons/icon.module.scss';
@@ -29,7 +29,7 @@ const Page: React.SFC = ({ children }) => {
             <InternalLink
               to="/customers"
               onClick={() => history.push('/customers')}
-              icon={<IconFemale className={styles.icon} />}
+              icon={<IconPerson className={styles.icon} />}
             >
               {t('common.sidebar.customers')}
             </InternalLink>,
